pyproject.toml
[tool.poetry]
name = "fastapi_auth_backend"
version = "0.1.0"
description = ""
authors = ["RezuanDzibov <rezuan.dzbov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.31"
alembic = "^1.7.5"
fastapi = "^0.73.0"
asyncpg = "^0.25.0"
uvicorn = "^0.17.1"
python-dotenv = "^0.19.2"
email-validator = "^1.1.3"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
python-multipart = "^0.0.5"
PyJWT = "^2.3.0"
emails = "^0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
